requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
google-api-python-client>=2.100.0
google-generativeai>=0.7.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
anthropic==0.7.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
